extract simulation engine helpers
